requirements.txt
requests==2.31.0
python-dotenv==1.0.1
eth-account==0.13.3
eth-utils==5.0.0
genlayer-test==0.28.1
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov>=4.1.0
